+++ tests/resources/js/emptyResource.js
@@ -1,4 +1,4 @@
-import { Component } from 'leto-modelizer-plugin-core';
+import TerraformComponent from 'src/models/TerraformComponent';
 import { getTerraformMetadata } from 'tests/resources/utils';
 
 const metadata = getTerraformMetadata(
@@ -10,7 +10,7 @@ metadata.parse();
 const instanceDefinition = metadata.pluginData.definitions.components.find(({ type }) => type === 'aws_instance');
 
 export default [
-  new Component({
+  new TerraformComponent({
     id: 'instance',
     name: null,
     path: 'new_file.tf',