+++ tests/resources/js/localReference.js
@@ -1,6 +1,5 @@
-import {
-  Component,
-} from 'leto-modelizer-plugin-core';
+import TerraformComponent from 'src/models/TerraformComponent';
+
 import TerraformComponentAttribute from 'src/models/TerraformComponentAttribute';
 import { getTerraformMetadata } from 'tests/resources/utils';
 import TerraformVariable from 'src/models/TerraformVariable';
@@ -25,7 +24,7 @@ export const localReference = [
 ];
 
 export const localComponentReference = [
-  new Component({
+  new TerraformComponent({
     name: null,
     id: 'database',
     path: 'new_file.tf',