+++ changelog.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [Unreleased]
+
+### Added
+
+- Create `TerraformComponent` class to override `createAttribute` function to create a new TerraformComponentAttribute instance.
+- Override `addComponent` function in `TerraformData` to create new component using TerraformComponent class.
+- Override `addLink`, `getLinkValue`, `removeLink`, `replaceLink`, `getReferenceValue` and `setReferenceValue` functions in `TerraformComponentAttribute` to handle link and reference attributes.
+
+### Changed
+
+- Update plugin-core to version 0.19.0.
+
 ## [0.4.1] - 2023/08/02
 
 ### Changed